FINAL/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
nltk>=3.6.0
requests>=2.25.0

# Vector search and embedding
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Lexical search
rank-bm25>=0.2.2

# Optional - For GPU acceleration
# faiss-gpu>=1.7.0
# torch>=1.10.0
# torchvision>=0.11.0

# LLM integration
huggingface-hub>=0.16.4

# Text processing
symspellpy>=6.7.0

# Data handling
ujson>=5.8.0
pyarrow>=12.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# CLI & UI 
streamlit==1.25.0
streamlit-chat>=0.1.0

# Web-related
flask>=2.0.0
flask-cors>=4.0.0

# Development tools
black>=23.7.0
isort>=5.12.0

# Added from the code block
plotly==5.15.0
streamlit-lottie==0.0.5 